+++ src/views/statusNode.ts
@@ -30,11 +30,11 @@ export class StatusNode extends ExplorerNode {
             children.push(new StatusUpstreamNode(status, 'ahead', this.context, this.git));
         }
 
-        if (status.files.length !== 0 || status.state.ahead && this.git.config.insiders) {
+        if (status.state.ahead || (status.files.length !== 0 && this.git.config.insiders)) {
             const range = status.state.ahead
                 ? `${status.upstream}..${status.branch}`
                 : undefined;
-            children.splice(0, 0, new StatusFilesNode(status, range, this.context, this.git));
+            children.push(new StatusFilesNode(status, range, this.context, this.git));
         }
 
         return children;
@@ -45,14 +45,16 @@ export class StatusNode extends ExplorerNode {
         if (status === undefined) return new TreeItem('No repo status');
 
         let hasChildren = false;
+        const hasWorkingChanges = status.files.length !== 0 && this.git.config.insiders;
         let label = '';
         let iconSuffix = '';
         if (status.upstream) {
             if (!status.state.ahead && !status.state.behind) {
-                label = `${status.branch} is up-to-date with ${status.upstream}`;
+                label = `${status.branch}${hasWorkingChanges ? ' has uncommitted changes and' : ''} is up-to-date with ${status.upstream}`;
             }
             else {
-                label = `${status.branch} is not up-to-date with ${status.upstream}`;
+                label = `${status.branch}${hasWorkingChanges ? ' has uncommitted changes and' : ''} is not up-to-date with ${status.upstream}`;
+
                 hasChildren = true;
                 if (status.state.ahead && status.state.behind) {
                     iconSuffix = '-yellow';
@@ -66,14 +68,10 @@ export class StatusNode extends ExplorerNode {
             }
         }
         else {
-            label = `${status.branch} is up-to-date`;
+            label = `${status.branch} ${hasWorkingChanges ? 'has uncommitted changes' : 'is clean'}`;
         }
 
-        if (this.git.config.insiders) {
-            hasChildren = hasChildren || status.files.length !== 0;
-        }
-
-        const item = new TreeItem(label, hasChildren ? TreeItemCollapsibleState.Expanded : TreeItemCollapsibleState.None);
+        const item = new TreeItem(label, (hasChildren || hasWorkingChanges) ? TreeItemCollapsibleState.Expanded : TreeItemCollapsibleState.None);
         item.contextValue = this.resourceType;
 
         item.iconPath = {